feat: add palindromes function
This function first set the inputted word to lowercase, checks to see if
each character is either a letter a-z or a number 0-9, then stores those
chosen characters into a variable declared as cleaned. Once the string
is completely checked it is then converted into an array, having its
order reversed, then converted back into a string. The original cleaned
string is then compared to the reversed string to see if it is a
palindrome. True if yes, false if no.